Split files into namespaced classes fro site health
